+++ cmd/flow_create.go
@@ -83,6 +83,9 @@ func runFlowCreate(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return cmdutil.HandleCommandError(cmd.ErrOrStderr(), err, jsonMode)
 	}
+	if err := cmdutil.CheckOK(resp); err != nil {
+		return cmdutil.HandleCommandError(cmd.ErrOrStderr(), err, jsonMode)
+	}
 
 	stdout := cmd.OutOrStdout()