+++ cmd/flow_delete.go
@@ -68,6 +68,9 @@ func runFlowDelete(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return cmdutil.HandleCommandError(cmd.ErrOrStderr(), err, jsonMode)
 	}
+	if err := cmdutil.CheckOK(resp); err != nil {
+		return cmdutil.HandleCommandError(cmd.ErrOrStderr(), err, jsonMode)
+	}
 
 	stdout := cmd.OutOrStdout()
 	if jsonMode {