+++ cmd/flow_deploy.go
@@ -79,6 +79,9 @@ func runFlowDeploy(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return cmdutil.HandleCommandError(cmd.ErrOrStderr(), err, jsonMode)
 	}
+	if err := cmdutil.CheckOK(resp); err != nil {
+		return cmdutil.HandleCommandError(cmd.ErrOrStderr(), err, jsonMode)
+	}
 
 	stdout := cmd.OutOrStdout()
 	if jsonMode {